// ui/src/gen/TimelineFactsRequest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TimelineToolStreamKey } from "./TimelineToolStreamKey";

export type TimelineFactsRequest = { expected_project_root: string, watched_tool_streams: Array<TimelineToolStreamKey>, watched_message_ids: number[], };